kraken/requirements.txt
python-dotenv
litestar
requests
questdb
psycopg